# docs/swagger.yaml
basePath: /api
definitions:
...
  models.AccountType:
    enum:
    - full
    - temporary
    type: string
    x-enum-comments:
      AccountTypeFull: Volledige account met app toegang
      AccountTypeTemporary: Tijdelijke registratie alleen voor event
    x-enum-descriptions:
    - Volledige account met app toegang
    - Tijdelijke registratie alleen voor event
    x-enum-varnames:
    - AccountTypeFull
    - AccountTypeTemporary
...
  models.Distance:
    properties:
      fund_amount:
        description: 'V26_08: Renamed from ''amount'''
        type: integer
      registration_fee:
        type: integer
      route:
        type: string
    type: object
  models.Event:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      end_time:
        type: string
      event_config:
        $ref: '#/definitions/models.EventConfig'
      geofences:
        items:
          $ref: '#/definitions/models.Geofence'
        type: array
      id:
        type: string
      is_active:
        type: boolean
      name:
        type: string
      start_time:
        type: string
      status:
        description: 'V27: Foreign key to lookup table (database column is ''status'')'
        type: string
      status_type:
        $ref: '#/definitions/models.EventStatusType'
      updated_at:
        type: string
    type: object
  models.EventConfig:
    additionalProperties: true
    type: object
...
  models.EventParticipantResponse:
    properties:
      check_in_time:
        type: string
      current_steps:
        description: Deprecated, kept for backwards compatibility
        type: integer
      distance_route:
        description: 'V26: Distance route'
        type: string
      event_id:
        type: string
      event_name:
        type: string
      finish_time:
        type: string
      id:
        type: string
      last_location_update:
        type: string
      participant_id:
        type: string
      participant_name:
        type: string
      participant_role:
        description: 'V26: Role name'
        type: string
      participant_role_description:
        description: 'V26: Role description'
        type: string
      registered_at:
        type: string
      start_time:
        type: string
      status:
        description: 'V27: Registration status'
        type: string
      status_description:
        description: 'V27: Status lookup description'
        type: string
      steps:
        description: Changed from CurrentSteps to match EventRegistration
        type: integer
      total_distance:
        type: number
      tracking_status:
        type: string
    type: object
  models.EventRegistration:
    properties:
      antwoorden_count:
        description: Antwoorden count voor tracking
        type: integer
      behandeld_door:
        type: string
      behandeld_op:
        type: string
      bijzonderheden:
        type: string
      check_in_time:
        type: string
      distance:
        $ref: '#/definitions/models.Distance'
      distance_route:
        description: 'V26: Database column is ''distance_route'''
        type: string
      email_verzonden:
        type: boolean
      email_verzonden_op:
        type: string
      event:
        $ref: '#/definitions/models.Event'
      event_id:
        type: string
      finish_time:
        type: string
      heeft_vervoer:
        description: 'V37: Transport vraag'
        type: boolean
      id:
        type: string
      last_location_update:
        type: string
      notities:
        type: string
      ondersteuning:
        type: string
      participant:
        allOf:
        - $ref: '#/definitions/models.Participant'
        description: '--- GORM Relaties ---'
      participant_id:
        type: string
      participant_role_name:
        description: 'V26: Database column is ''participant_role_name'''
        type: string
      registered_at:
        type: string
      registration_status:
        $ref: '#/definitions/models.RegistrationStatusType'
      role:
        $ref: '#/definitions/models.ParticipantRole'
      start_time:
        type: string
      status:
        description: |-
          --- Foreign Keys (V26 & V27) ---
          V27: Database column is 'status' (renamed from status_key)
        type: string
      steps:
        description: '--- Velden verplaatst van ''Aanmelding'' (nu Participant) ---'
        type: integer
      terms:
        description: 'V37: Terms veld behouden in event_registrations'
        type: boolean
      test_mode:
        type: boolean
      total_distance:
        type: number
      tracking_status:
        description: GPS Tracking velden (V23 event_participants)
        type: string
    type: object
...
  models.EventStatusType:
    properties:
      description:
        type: string
      status:
        description: 'Database column: status'
        type: string
    type: object
...
  models.Geofence:
    properties:
      lat:
        description: Latitude
        type: number
      long:
        description: Longitude
        type: number
      name:
        type: string
      radius:
        description: Radius in meters
        type: number
      type:
        description: '"start", "checkpoint", "finish"'
        type: string
    type: object
  models.Participant:
    properties:
      account_type:
        allOf:
        - $ref: '#/definitions/models.AccountType'
        description: 'V30: Duaal registratiesysteem velden'
      antwoorden:
        description: Relatie met antwoorden
        items:
          $ref: '#/definitions/models.ParticipantAntwoord'
        type: array
      created_at:
        type: string
      email:
        type: string
      event_registrations:
        description: Relatie met event registraties
        items:
          $ref: '#/definitions/models.EventRegistration'
        type: array
      gebruiker_id:
        type: string
      has_app_access:
        type: boolean
      id:
        type: string
      naam:
        type: string
      registration_year:
        description: 'Voor temporary accounts: evenementjaar'
        type: integer
      telefoon:
        type: string
      terms:
        description: '''terms'' bleef op Participant (V28)'
        type: boolean
      test_mode:
        description: Dit veld ontbrak nog (uit V1)
        type: boolean
      updated_at:
        type: string
      upgraded_at:
        type: string
      upgraded_to_gebruiker_id:
        type: string
    type: object
  models.ParticipantAntwoord:
    properties:
      email_verzonden:
        type: boolean
      id:
        type: string
      participant_id:
        description: Hernoemd
        type: string
      tekst:
        type: string
      verzonden_door:
        type: string
      verzonden_op:
        type: string
    type: object
  models.ParticipantRole:
    properties:
      description:
        type: string
      is_active:
        description: Required by repository queries
        type: boolean
      name:
        type: string
    type: object
...
  models.RegistrationStatusType:
    properties:
      description:
        type: string
      status:
        description: 'Database column: status'
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: info@dekoninklijkeloop.nl
    name: API Support
    url: http://www.dekoninklijkeloop.nl
  description: API voor de De Koninklijke Loop email service - inclusief registratie,
    deelnemersbeheer, CMS en meer
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DKL Email Service API
  version: "1.0"
paths:
...
  /api/events/{id}/participants:
    get:
      consumes:
      - application/json
      description: Haalt lijst van participants voor een event op
      parameters:
      - description: Event ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.EventParticipantResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Event Participants ophalen
      tags:
      - Events
...
  /api/participant:
    get:
      consumes:
      - application/json
      description: Haalt een gepagineerde lijst van participants (personen) op
      parameters:
      - description: Aantal resultaten per pagina (standaard 10)
        in: query
        name: limit
        type: integer
      - description: Offset voor paginering (standaard 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Participant'
            type: array
      security:
      - BearerAuth: []
      summary: Lijst van participants ophalen
      tags:
      - Participant
  /api/participant/{id}:
    delete:
      consumes:
      - application/json
      description: Verwijdert een participant en bijbehorende antwoorden (en via CASCADE
        ook EventRegistrations)
      parameters:
      - description: Participant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Participant verwijderen
      tags:
      - Participant
    get:
      consumes:
      - application/json
      description: Haalt de details van een specifieke participant op, inclusief antwoorden
      parameters:
      - description: Participant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Participant'
      security:
      - BearerAuth: []
      summary: Details van een specifieke participant ophalen
      tags:
      - Participant
  /api/participant/{id}/antwoord:
    post:
      consumes:
      - application/json
      description: Voegt een nieuw antwoord toe aan een participant
      parameters:
      - description: Participant ID
        in: path
        name: id
        required: true
        type: string
      - description: Antwoord gegevens
        in: body
        name: antwoord
        required: true
        schema:
          $ref: '#/definitions/models.ParticipantAntwoord'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ParticipantAntwoord'
      security:
      - BearerAuth: []
      summary: Antwoord toevoegen aan participant
      tags:
      - Participant
...
  /api/participant/emails:
    get:
      consumes:
      - application/json
      description: Haalt een lijst van alle participant emails en systeem emails op
        voor admin gebruik
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              ' all_emails':
                type: array
              ' system_emails':
                type: array
              participant_emails:
                type: array
            type: object
      security:
      - BearerAuth: []
      summary: Alle emails ophalen (participants + systeem)
      tags:
      - Participant
...
securityDefinitions:
  BearerAuth:
    description: JWT Bearer token voor geauthenticeerde endpoints
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"